third_party/blink/web_tests/fast/frames/frame-inherit-noresize-from-frameset.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/frame-programmatic-resize.js"></script>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFrames();
}

window.onload = runTests;

function runTests()
{
    if (!window.eventSender)
        return; // Cannot run this test without DRT.

    description("This tests that a frame inherits the noresize attribute of its parent frameset by default. Note, this behavior doesn't seem to conform to any W3C spec or MSDN documentation. See WebKit Bug 57604.");
    setTestFrameById("testFrame");
    shouldDisallowFrameResize();
}
</script>
</head>
<frameset cols="200,*" border="10" noresize>
    <frame id="testFrame" name="testFrame" src="about:blank">
    <frame id="results" name="results" src="data:text/html,<p id='description'>This test can only be run using DRT.</p><pre id='console'></pre><script>window.onmessage = e => { if (e.data.type === 'description') { document.getElementById('description').textContent = e.data.message; } else { document.getElementById('console').appendChild(document.createTextNode(e.data.message + '\n')); } }</script>">
</frameset>
</html>
