third_party/blink/web_tests/platform/generic/fast/frames/frame-inherit-noresize-from-frameset-expected.txt



--------
Frame: 'testFrame'
--------


--------
Frame: 'results'
--------
This tests that a frame inherits the noresize attribute of its parent frameset by default. Note, this behavior doesn't seem to conform to any W3C spec or MSDN documentation. See WebKit Bug 57604.

PASS document.getElementById("testFrame").getBoundingClientRect().width is 200
PASS document.getElementById("testFrame").getBoundingClientRect().width is 210
